Do not disable influxdb in systemd during an upgrade
RPM runs the %postun action of the old package after it runs the %post
action of the new package. Dpkg seems to call postrm of the old package
before it calls postinst of the new package, so there isn't a problem
there.

If we are on a Red Hat machine and are upgrading the package, exit the
post-uninstall script instead or running it.